+++ openfisca_core/base_functions.py
@@ -1,0 +1,99 @@
+import numpy as np
+
+def permanent_default_value(formula, simulation, period):
+    if formula.function is not None:
+        return formula.function(simulation, period)
+    holder = formula.holder
+    column = holder.column
+    array = np.empty(holder.entity.count, dtype = column.dtype)
+    array.fill(column.default)
+    return period, array
+
+
+def requested_period_added_value(formula, simulation, period):
+    # This formula is used for variables that can be added to match requested period.
+    holder = formula.holder
+    column = holder.column
+    period_size = period.size
+    period_unit = period.unit
+    if holder._array_by_period is not None and (period_size > 1 or period_unit == u'year'):
+        after_instant = period.start.offset(period_size, period_unit)
+        if period_size > 1:
+            array = formula.zeros(dtype = column.dtype)
+            sub_period = period.start.period(period_unit)
+            while sub_period.start < after_instant:
+                sub_array = holder._array_by_period.get(sub_period)
+                if sub_array is None:
+                    array = None
+                    break
+                array += sub_array
+                sub_period = sub_period.offset(1)
+            if array is not None:
+                return period, array
+        if period_unit == u'year':
+            array = formula.zeros(dtype = column.dtype)
+            month = period.start.period(u'month')
+            while month.start < after_instant:
+                month_array = holder._array_by_period.get(month)
+                if month_array is None:
+                    array = None
+                    break
+                array += month_array
+                month = month.offset(1)
+            if array is not None:
+                return period, array
+    if formula.function is not None:
+        return formula.function(simulation, period)
+    array = np.empty(holder.entity.count, dtype = column.dtype)
+    array.fill(column.default)
+    return period, array
+
+
+def requested_period_default_value(formula, simulation, period):
+    if formula.function is not None:
+        return formula.function(simulation, period)
+    holder = formula.holder
+    column = holder.column
+    array = np.empty(holder.entity.count, dtype = column.dtype)
+    array.fill(column.default)
+    return period, array
+
+
+def requested_period_default_value_neutralized(formula, simulation, period):
+    holder = formula.holder
+    column = holder.column
+    array = np.empty(holder.entity.count, dtype = column.dtype)
+    array.fill(column.default)
+    return period, array
+
+
+def requested_period_last_value(formula, simulation, period):
+    # This formula is used for variables that are constants between events and period size independent.
+    # It returns the latest known value for the requested period.
+    holder = formula.holder
+    if holder._array_by_period is not None:
+        for last_period, last_array in sorted(holder._array_by_period.iteritems(), reverse = True):
+            if last_period.start <= period.start and (formula.function is None or last_period.stop >= period.stop):
+                return period, last_array
+    if formula.function is not None:
+        return formula.function(simulation, period)
+    column = holder.column
+    array = np.empty(holder.entity.count, dtype = column.dtype)
+    array.fill(column.default)
+    return period, array
+
+
+def last_duration_last_value(formula, simulation, period):
+    # This formula is used for variables that are constants between events but are period size dependent.
+    # It returns the latest known value for the requested start of period but with the last period size.
+    holder = formula.holder
+    if holder._array_by_period is not None:
+        for last_period, last_array in sorted(holder._array_by_period.iteritems(), reverse = True):
+            if last_period.start <= period.start and (formula.function is None or last_period.stop >= period.stop):
+                return periods.Period((last_period[0], period.start, last_period[2])), last_array
+    if formula.function is not None:
+        return formula.function(simulation, period)
+    column = holder.column
+    array = np.empty(holder.entity.count, dtype = column.dtype)
+    array.fill(column.default)
+    return period, array
